Move debug to new file, rename stuff.
