+++ functions.py
@@ -3,8 +3,9 @@ import cv2
 import numpy as np
 import matplotlib.pyplot as plt
 from sklearn.cluster import DBSCAN
-import matplotlib.colors as mcolors
-from matplotlib.patches import Patch
+
+from debug import _debug_show_clusters, _debug_show_cluster_normalized, _debug_plot_histogram, _debug_show_image
+
 import math
 
 from pathlib import Path
@@ -66,11 +67,10 @@ def get_percentile(channel_img, percentile):
 #   Global thresholding: Threshold percentile
 #   DBSCAN: Epsilon and min_samples
 # Done by computing each combination and using best results of objective function.
-def determine_best_parameters(
+def compute_threshold(
     channel_img: np.ndarray,
     thresh_range: Tuple[float, float],
     ):
-    scores = []
     for percentile in range(thresh_range[0], thresh_range[1], 1):
         percentile /= 100
         print("Testing threshold " + str(percentile))
@@ -78,11 +78,7 @@ def determine_best_parameters(
         cluster_mask = find_clusters(thresholded_img)
         # _debug_show_clusters(cluster_mask)
         process_clusters(channel_img, cluster_mask)
-        scores.append(objective(cluster_mask))
-    print(scores)
-    # Will be for looking at results of objective function.
-    # plt.scatter(np.arange(thresh_range[0], thresh_range[1], 1), scores)
-    # plt.show()
+
 
 # Use the cluster mask to extract what was detected to be each cluster.
 # After isolating each cluster, normalize the brightnesses to 0-255.
@@ -104,17 +100,18 @@ def process_clusters(channel_img, cluster_mask):
         masked = np.full_like(channel_img, fill_value=cluster_min)
         masked[mask] = channel_img[mask]
         
-        # normalize the brightnesses to be in 0-255.
+        # The cluster points are copied, now normalize the brightnesses to be in 0-255.
         remapped = remap_values(masked, cluster_min, cluster_max, 0, 255)
         cluster_cilia.append(isolate_bright_spots(remapped))
+        
         #_debug_show_cluster_normalized(remapped, cluster)
     
-    #Combine masks
+    #Stack the masks together for the final thresholded image.
     stacked = np.zeros(dtype=np.uint8, shape=channel_img.shape)
     for cluster_cilia_mask in cluster_cilia:
         stacked = cv2.bitwise_or(stacked, cluster_cilia_mask)
-    plt.imshow(stacked, interpolation='none')
-    plt.show()
+    _debug_show_image(stacked, "Final step: thresholded clusters stacked together")
+    return stacked
 
 # Remap values of a 2d array from one range to another. min1 max1 to min2 max2.
 # might result in floating point values?
@@ -129,6 +126,10 @@ def remap_values(img, min1, max1, min2, max2):
             img[i,j] = remapped
     return img
 
+
+# Take an isolated cluster with ranges from 0-255, and isolate the bright spots
+# which are hopefully at this point made up mostly of the features we want.
+# Currently mostly tuned for cilia.
 def isolate_bright_spots(img):
     img = np.uint8(img)
     thresh = cv2.adaptiveThreshold(
@@ -141,76 +142,27 @@ def isolate_bright_spots(img):
     )
     kernel_size = (3,3) #May depend on zoom
     kernel = cv2.getStructuringElement(cv2.MORPH_ELLIPSE, kernel_size)
+    
+    # Get rid of tiny artifacts
     opened_img = cv2.morphologyEx(
         thresh,
         cv2.MORPH_OPEN,
         kernel,
     )
-    # plt.imshow(thresh, interpolation='none')
-    # plt.show()
-    # plt.imshow(opened_img, interpolation='none')
-    # plt.show()
+    # _debug_show_image(thresh)
+    # _debug_show_image(opened_img)
     return opened_img
-
-# DEBUG. Visualize computed clusters.
-def _debug_show_clusters(cluster_mask):
-    cluster_ids = sorted(list(set(cluster_mask.flatten()) - {-2, -1}))
-    num_clusters = len(cluster_ids)
-    print("Num clusters: " + str(num_clusters))
-    
-    values = set(cluster_mask.flatten())
-    counts = {-2: 0, -1: 0}
-    for value in values:
-        count = (cluster_mask == value).sum().item()
-        counts[value.item()] = count
-    
-    # Set up visualization
-    cmap = plt.get_cmap('tab20')
-    norm = mcolors.Normalize(vmin=-2, vmax=cluster_mask.max())
-
-    plt.imshow(cluster_mask, cmap=cmap, norm=norm, interpolation='none')
-    plt.suptitle("clusters")
-    # plt.set_xlabel(f"num clusters={num_clusters}")
-
-    legend_colors = [cmap(norm(val)) for val in values]
-    patches = [Patch(color=color, label=f"cluster {val}: {counts[val]}") for val, color in zip(values, legend_colors)]
-    plt.legend(handles=patches, bbox_to_anchor=(1.85,1.0))
-    plt.show()
-
-#Used for showing each cluster isolated from everything else.
-# Auto zooms to important area.
-def _debug_show_cluster_normalized(masked_remapped_cluster, number):
-    plt.imshow(masked_remapped_cluster, interpolation='none')
-    plt.suptitle("cluster " + str(number))
-    plt.show()
-
-# Display the pixel brightnesses of an channel image as a histogram.
-# Use start if you want to start from a specific brightness
-def _debug_plot_histogram(channel_img, start=0):
-    plt.xlabel("Pixel brightness")
-    plt.ylabel("Pixel count")
-    
-    max_val = np.max(channel_img)
-    plt.plot(cv2.calcHist([channel_img], [0], None, [max_val], [start,max_val]))
-    plt.show()
-    
-
-def objective(cluster_mask):
-    # FIXME
-    num_clusters = cluster_mask.max() + 1
-    noise_cluster_size = (cluster_mask == -1).sum()
-    return num_clusters - noise_cluster_size
-
 
 def find_clusters(
     channel_img: np.ndarray,
     eps=25, # Epsilon should be a function of the zoom of the image as well. For now, just some number I pulled out of nowhere.
+    samples=125
 ) -> np.ndarray:
 
     points = np.column_stack(np.where(channel_img > 0))
 
     # find clusters using DBSCAN
-    dbscan = DBSCAN(eps=eps, min_samples=125) # For now, just some number I pulled out of nowhere. about this dense is good.
+    dbscan = DBSCAN(eps=eps, min_samples=samples) # For now, just some number I pulled out of nowhere. about this dense is good.
     labels = dbscan.fit_predict(points)
 
     # mark each cluster
